Extract mega sort funciton and add useful comments for editable entity page
