+++ apps/web/modules/components/entity/editable-entity-page.tsx
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-import { SYSTEM_IDS } from '@geogenesis/ids';
 import Head from 'next/head';
 import { useActionsStore } from '~/modules/action';
 import { Button, SquareButton } from '~/modules/design-system/button';
@@ -15,6 +14,7 @@ import { EntityAutocompleteDialog } from './autocomplete/entity-autocomplete';
 import { EntityTextAutocomplete } from './autocomplete/entity-text-autocomplete';
 import { CopyIdButton } from './copy-id';
 import { useEditEvents } from './edit-events';
+import { sortEditableEntityPageTriples } from './editable-entity-page-utils';
 import { NumberField, PlaceholderField, StringField } from './editable-fields';
 import { TripleTypeDropdown } from './triple-type-dropdown';
 
@@ -254,54 +254,11 @@ function EntityAttributes({
     return notHidden && notInTriples;
   });
 
-  const filteredTriples = [...triples, ...visibleSchemaTriples];
+  const visibleTriples = [...triples, ...visibleSchemaTriples];
 
   const entityValueTriples = triples.filter(triple => triple.value.type === 'entity');
 
-  const schemaAttributeIds = schemaTriples.map(schemaTriple => schemaTriple.attributeId);
-
-  const sortedTriples = filteredTriples.sort((tripleA, tripleB) => {
-    /* Sort order goes Name -> Description -> Types -> Placeholders (Empty or modified) -> New triples */
-
-    const { attributeId: attributeIdA } = tripleA;
-    const { attributeId: attributeIdB } = tripleB;
-
-    const aName = attributeIdA === SYSTEM_IDS.NAME;
-    const bName = attributeIdB === SYSTEM_IDS.NAME;
-    const aDescription = attributeIdA === SYSTEM_IDS.DESCRIPTION;
-    const bDescription = attributeIdB === SYSTEM_IDS.DESCRIPTION;
-    const aTypes = attributeIdA === SYSTEM_IDS.TYPES;
-    const bTypes = attributeIdB === SYSTEM_IDS.TYPES;
-
-    const aIndex = schemaAttributeIds.indexOf(attributeIdA);
-    const bIndex = schemaAttributeIds.indexOf(attributeIdB);
-
-    const aInSchema = schemaAttributeIds.includes(attributeIdA);
-    const bInSchema = schemaAttributeIds.includes(attributeIdB);
-
-    if (aName && !bName) return -1;
-    if (!aName && bName) return 1;
-
-    if (aDescription && !bDescription) return -1;
-    if (!aDescription && bDescription) return 1;
-
-    if (aTypes && !bTypes) return -1;
-    if (!aTypes && bTypes) return 1;
-
-    if (aInSchema && !bInSchema) {
-      return -1;
-    }
-
-    if (!aInSchema && bInSchema) {
-      return 1;
-    }
-
-    if (aInSchema && bInSchema) {
-      return aIndex - bIndex;
-    }
-
-    return 0;
-  });
+  const sortedTriples = sortEditableEntityPageTriples(visibleTriples, schemaTriples);
 
   const groupedTriples = groupBy(sortedTriples, triple => triple.attributeId);
   const attributeIds = Object.keys(groupedTriples);
